+++ src/components/Typography/index.ts
@@ -1,3 +1,3 @@
-// Typography from copy-ui (version: 0.1.11, date: 2025-12-12).
+// Typography from copy-ui (version: 0.1.12, date: 2026-02-18).
 
 export { Typography } from "./typography";