+++ src/components/Typography/typography-code.tsx
@@ -1,7 +1,6 @@
 import classNames from "classnames";
 import type { ComponentProps, FC } from "react";
 
-import { CodeBlock } from "@/components/CodeBlock";
 import { useTheme } from "@/components/CopyUiProvider";
 
 import styles from "./typography-code.module.css";
@@ -33,24 +32,7 @@ const TypographyCode: FC<TypographyCodeProps> = (props) => {
   );
 };
 
-// Typography.CodeBlock
-// =============================================================================
-
-type TypographyCodeBlockProps = ComponentProps<typeof CodeBlock>;
-
-const TypographyCodeBlock: FC<TypographyCodeBlockProps> = (props) => {
-  const { className, ...rest } = props;
-
-  return (
-    <CodeBlock
-      scrollAreaMaxHeight="calc(50vh + 10rem)"
-      className={classNames(styles.codeBlock, className)}
-      {...rest}
-    />
-  );
-};
-
 // Export
 // =============================================================================
 
-export { TypographyCode, TypographyCodeBlock };
+export { TypographyCode };